fixed tests for existing conv & linear decoders
